+++ apps/io-services-cms-webapp/src/webservice/controllers/get-services.ts
@@ -1,8 +1,9 @@
 import { SubscriptionContract } from "@azure/arm-apimanagement";
-import { Context } from "@azure/functions";
+import { InvocationContext } from "@azure/functions";
 import { ApimUtils } from "@io-services-cms/external-clients";
 import { ServiceLifecycle } from "@io-services-cms/models";
 import { SubscriptionCIDRsModel } from "@pagopa/io-functions-commons/dist/src/models/subscription_cidrs";
+import { wrapHandlerV4 } from "@pagopa/io-functions-commons/dist/src/utils/azure-functions-v4-express-adapter";
 import {
   AzureApiAuthMiddleware,
   IAzureApiAuthorization,
@@ -11,10 +12,6 @@ import {
 import { ClientIpMiddleware } from "@pagopa/io-functions-commons/dist/src/utils/middlewares/client_ip_middleware";
 import { ContextMiddleware } from "@pagopa/io-functions-commons/dist/src/utils/middlewares/context_middleware";
 import { OptionalQueryParamMiddleware } from "@pagopa/io-functions-commons/dist/src/utils/middlewares/optional_query_param";
-import {
-  withRequestMiddlewares,
-  wrapRequestHandler,
-} from "@pagopa/io-functions-commons/dist/src/utils/request_middleware";
 import {
   IWithinRangeIntegerTag,
   IntegerFromString,
@@ -51,7 +48,7 @@ import { ErrorResponseTypes, getLogger } from "../../utils/logger";
 const logPrefix = "GetServicesHandler";
 
 type GetServicesHandler = (
-  context: Context,
+  context: InvocationContext,
   auth: IAzureApiAuthorization,
   limit: O.Option<number>,
   offset: O.Option<number>,
@@ -245,8 +242,8 @@ export const makeGetServicesHandler =
 
 export const applyRequestMiddelwares =
   (config: IConfig, subscriptionCIDRsModel: SubscriptionCIDRsModel) =>
-  (handler: GetServicesHandler): ReturnType<typeof wrapRequestHandler> => {
-    const middlewaresWrap = withRequestMiddlewares(
+  (handler: GetServicesHandler): ReturnType<typeof wrapHandlerV4> => {
+    const middlewares = [
       // extract the client IP from the request
       ClientIpMiddleware,
       // check manage key
@@ -275,20 +272,18 @@ export const applyRequestMiddelwares =
         IntegerFromString.pipe(NonNegativeInteger),
       ),
       SelfcareUserGroupsMiddleware(),
-    );
-    return wrapRequestHandler(
-      middlewaresWrap((...args) =>
-        pipe(
-          args,
-          ([clientIp, userAttributesManage, ...rest]) =>
-            pipe(
-              checkSourceIp(clientIp, userAttributesManage.authorizedCIDRs),
-              E.map((_) => rest),
-            ),
-          TE.fromEither,
-          TE.chainW((args) => handler(...args)),
-          TE.toUnion,
-        )(),
-      ),
+    ] as const;
+    return wrapHandlerV4(middlewares, (...args) =>
+      pipe(
+        args,
+        ([clientIp, userAttributesManage, ...rest]) =>
+          pipe(
+            checkSourceIp(clientIp, userAttributesManage.authorizedCIDRs),
+            E.map((_) => rest),
+          ),
+        TE.fromEither,
+        TE.chainW((args) => handler(...args)),
+        TE.toUnion,
+      )(),
     );
   };