+++ apps/io-services-cms-webapp/CHANGELOG.md
@@ -1,5 +1,20 @@
 # io-services-cms-webapp
 
+## 2.0.1
+
+### Patch Changes
+
+- e5411807: upgrade @pagopa/io-functions-commons version to v30.0.0
+- Updated dependencies [e5411807]
+  - @io-services-cms/models@1.29.0
+  - @io-services-cms/external-clients@1.9.0
+
+## 2.0.0
+
+### Major Changes
+
+- 73d9dd9b: upgrade Az Fn to PMv4 and upgrade node version to 22
+
 ## 1.42.12
 
 ### Patch Changes